add_custom_detectors_to_metadata_integration before multiprocess
