+++ checkov/main.py
@@ -8,6 +8,7 @@ import os
 import shutil
 import signal
 import sys
+from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING
 
@@ -936,5 +937,10 @@ class Checkov:
 
 
 if __name__ == '__main__':
+    start_time = datetime.now()
     ckv = Checkov()
-    sys.exit(ckv.run())
+    end_time = datetime.now()
+    ckv.run()
+    logging.debug(f'total time = {end_time - start_time}')
+    sys.exit()
+
